refactor: rename format.ts > Date.ts
